Move disc search and filters to separate component, add another header nav item with type 'Approach', adjust filters
