+++ src/components/header.tsx
@@ -71,7 +71,7 @@ export default function Header() {
                       </li>
                       <li>
                         <Link
-                          href={`${Page.Discs}?type=distance`}
+                          href={`${Page.Discs}?type=Distance+Driver`}
                           className="hover:bg-accent hover:text-accent-foreground focus:bg-accent focus:text-accent-foreground block space-y-1 rounded-md p-3 leading-none no-underline transition-colors outline-none select-none"
                         >
                           <div className="text-sm leading-none font-medium">
@@ -84,7 +84,7 @@ export default function Header() {
                       </li>
                       <li>
                         <Link
-                          href={`${Page.Discs}?type=control`}
+                          href={`${Page.Discs}?type=Control+Driver`}
                           className="hover:bg-accent hover:text-accent-foreground focus:bg-accent focus:text-accent-foreground block space-y-1 rounded-md p-3 leading-none no-underline transition-colors outline-none select-none"
                         >
                           <div className="text-sm leading-none font-medium">
@@ -97,20 +97,33 @@ export default function Header() {
                       </li>
                       <li>
                         <Link
-                          href={`${Page.Discs}?type=midrange`}
+                          href={`${Page.Discs}?type=Midrange`}
                           className="hover:bg-accent hover:text-accent-foreground focus:bg-accent focus:text-accent-foreground block space-y-1 rounded-md p-3 leading-none no-underline transition-colors outline-none select-none"
                         >
                           <div className="text-sm leading-none font-medium">
                             Midranges
                           </div>
                           <p className="text-muted-foreground line-clamp-2 text-sm leading-snug">
-                            Versatile discs for approach shots
+                            Versatile discs for mid-range shots
                           </p>
                         </Link>
                       </li>
                       <li>
                         <Link
-                          href={`${Page.Discs}?type=putter`}
+                          href={`${Page.Discs}?type=Approach`}
+                          className="hover:bg-accent hover:text-accent-foreground focus:bg-accent focus:text-accent-foreground block space-y-1 rounded-md p-3 leading-none no-underline transition-colors outline-none select-none"
+                        >
+                          <div className="text-sm leading-none font-medium">
+                            Approach Discs
+                          </div>
+                          <p className="text-muted-foreground line-clamp-2 text-sm leading-snug">
+                            Perfect for approaching the basket
+                          </p>
+                        </Link>
+                      </li>
+                      <li>
+                        <Link
+                          href={`${Page.Discs}?type=Putter`}
                           className="hover:bg-accent hover:text-accent-foreground focus:bg-accent focus:text-accent-foreground block space-y-1 rounded-md p-3 leading-none no-underline transition-colors outline-none select-none"
                         >
                           <div className="text-sm leading-none font-medium">
