+++ src/constants/filters.ts
@@ -1,0 +1,4 @@
+export const DEFAULT_PAGE = 1;
+export const DEFAULT_PER_PAGE = 20;
+export const DEFAULT_SPEED_RANGE = [1, 14];
+export const DEBOUCE_DELAY = 500;
